chore: create dedicated StreamHandler class
